TST: move and consolidate Geometry __eq__ tests in geometry/test_equality.py
